# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "API for managing car rentals",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Récupérer une voiture spécifique",
                "operationId": "27c826192271dae5e6e7246be568b4ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voiture récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voiture non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "required": [
                    "id",
                    "model",
                    "daily_rate"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "model": {
                        "type": "string",
                        "example": "Toyota Corolla"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://example.com/car.jpg"
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "daily_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    }
                },
                "type": "object"
            }
        }
    },
...
}